SmallMap: remove SetUnique functions
The intent of those functions was to flag cases where information is
unintentionally being overwritten. However, the expected usage of these
maps is local construction of error info, followed by being frozen, so
the small set of keys should be easy to keep unique. Simplicity of the
API is more important.