Remove domain registration
The requirement to register domains before use was intended to provide
checks for e.g. misspelled domain names, but introduced more complexity
that it was worth, especially given that the error would only be
detected at runtime (in an error path) anyway.

By removing the domain registry, there is now no static/global data in
the library, meaning that it can be used as a shared library if desired.

It should be noted that client systems can implement domain name
validation if they desire, and in any case following good coding
practice should in most cases prevent errors.